+++ index.html
@@ -211,6 +211,6 @@
         <img class="icon" src="assets/img/icon-search.svg" alt="icon-search">
         <img class="icon" src="assets/img/icon-bell.svg" alt="icon-bell">
     </footer>
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
 </body>
 </html>
